Some tweak edits made to the Search page to work
Removed the search bar from the search page. Changed the search span in header
to a button so it can submit a POST request. Edited some of the PHP for the search function.
Added "no results" text if no matches are found.
Made the CSS style the same as homepage.